Add Synchroniser struct
Adds `Synchroniser` struct responsible for performing synchronisation
and exposing an error object which neatly encapsulates the various
ways in which sync can fail. Also fixes a few translation imports
which were causing compilation warnings.
