+++ src/util/sync/mod.rs
@@ -3,6 +3,7 @@ mod connection;
 mod credentials;
 mod remote;
 mod server;
+mod synchroniser;
 mod translation;
 
 pub use central::CentralSyncBatch;
@@ -13,24 +14,17 @@ pub use remote::{
     RemoteSyncRecordData,
 };
 pub use server::SyncServer;
+pub use synchroniser::Synchroniser;
 
-use crate::{
-    database::repository::{
-        CentralSyncBufferRepository, CentralSyncCursorRepository, SyncRepository,
-    },
-    server::data::RepositoryRegistry,
-};
+use crate::server::data::RepositoryRegistry;
 
-use actix_web::web::Data;
 use log::info;
 use tokio::{
     sync::mpsc::{self, error as mpsc_error, Receiver as MpscReceiver, Sender as MpscSender},
     time::{self, Duration, Interval},
 };
 
-use self::translation::{import_sync_records, TRANSLATION_RECORDS};
-
-pub fn get_sync_actors(connection: SyncConnection) -> (SyncSenderActor, SyncReceiverActor) {
+pub fn get_sync_actors() -> (SyncSenderActor, SyncReceiverActor) {
     // We use a single-element channel so that we can only have one sync pending at a time.
     // We consume this at the *start* of sync, so we could schedule a sync while syncing.
     // Worst-case scenario, we produce an infinite stream of sync instructions and always go
@@ -38,10 +32,7 @@ pub fn get_sync_actors(connection: SyncConnection) -> (SyncSenderActor, SyncRece
     let (sender, receiver) = mpsc::channel(1);
 
     let sync_sender = SyncSenderActor { sender };
-    let sync_receiver = SyncReceiverActor {
-        connection,
-        receiver,
-    };
+    let sync_receiver = SyncReceiverActor { receiver };
 
     (sync_sender, sync_receiver)
 }
@@ -77,224 +68,21 @@ impl SyncSenderActor {
     }
 }
 pub struct SyncReceiverActor {
-    connection: SyncConnection,
     receiver: MpscReceiver<()>,
 }
 
 #[allow(unused_assignments)]
 impl SyncReceiverActor {
-    pub async fn pull_central_records(&mut self, repositories: &RepositoryRegistry) {
-        let central_sync_cursor_repository: &CentralSyncCursorRepository =
-            repositories.get::<CentralSyncCursorRepository>();
-
-        let central_sync_buffer_repository: &CentralSyncBufferRepository =
-            repositories.get::<CentralSyncBufferRepository>();
-
-        let mut cursor: u32 = central_sync_cursor_repository
-            .get_cursor()
-            .await
-            .unwrap_or_else(|_| {
-                info!("Initialising new central sync cursor...");
-                0
-            });
-
-        // Arbitrary batch size.
-        const BATCH_SIZE: u32 = 500;
-
-        loop {
-            info!("Requesting central sync data...");
-            let sync_batch: CentralSyncBatch = self
-                .connection
-                .central_records(cursor, BATCH_SIZE)
-                .await
-                .expect("Failed to pull central sync records");
-            info!("Received central sync response");
-
-            if let Some(central_sync_records) = sync_batch.data {
-                for central_sync_record in central_sync_records {
-                    central_sync_buffer_repository
-                        .insert_one_and_update_cursor(&central_sync_record)
-                        .await
-                        .expect("Failed to insert central sync record into sync buffer");
-                }
-            }
-
-            cursor = central_sync_cursor_repository
-                .get_cursor()
-                .await
-                .expect("Failed to load central sync cursor");
-
-            if cursor >= sync_batch.max_cursor - 1 {
-                info!("All central sync records pulled successfully");
-                break;
-            }
-        }
-    }
-
-    // Hacky method for pulling from sync_queue.
-    pub async fn pull_remote_records(&mut self) -> Vec<RemoteSyncRecord> {
-        // TODO: only initialize on initial sync.
-        info!("Sending initialize request...");
-        let mut sync_batch: RemoteSyncBatch = self
-            .connection
-            .initialize()
-            .await
-            .expect("Failed to initialize remote sync records");
-        info!("Received initialize response");
-
-        let mut records: Vec<RemoteSyncRecord> = Vec::new();
-        while sync_batch.queue_length > 0 {
-            info!("Sending remote sync request...");
-            sync_batch = self
-                .connection
-                .remote_records()
-                .await
-                .expect("Failed to pull remote sync records");
-            info!("Received remote sync response");
-
-            // TODO: acknowledge after integration.
-            if let Some(data) = sync_batch.data {
-                records.append(&mut data.clone());
-                info!("Acknowledging remote sync records...");
-                self.connection
-                    .acknowledge_records(&records)
-                    .await
-                    .expect("Failed to acknowledge remote sync records");
-                info!("Acknowledged remote sync records");
-            }
-        }
-
-        records
-    }
-
-    pub fn integrate_remote_records(&self, records: Vec<RemoteSyncRecord>) {
-        records.iter().for_each(|record| {
-            info!("Integrated remote sync record {}", record.sync_id);
-        });
-    }
-
-    async fn integrate_central_records(
-        &self,
-        repositories: &RepositoryRegistry,
-    ) -> Result<(), String> {
-        let central_sync_buffer_repository: &CentralSyncBufferRepository =
-            repositories.get::<CentralSyncBufferRepository>();
-        let sync_session = repositories
-            .get::<SyncRepository>()
-            .new_sync_session()
-            .await
-            .unwrap();
-        for table_name in TRANSLATION_RECORDS {
-            let buffer_rows = central_sync_buffer_repository
-                .get_sync_entries(table_name)
-                .await
-                .map_err(|_| "Failed to read central sync entries".to_string())?;
-            import_sync_records(&sync_session, repositories, &buffer_rows).await?;
-        }
-        central_sync_buffer_repository
-            .remove_all()
-            .await
-            .map_err(|_| "Failed to empty central sync entries".to_string())?;
-        Ok(())
-    }
-
     // Listen for incoming sync messages.
-    pub async fn listen(&mut self, repositories: Data<RepositoryRegistry>) {
+    pub async fn listen(&mut self, synchroniser: &mut Synchroniser, registry: &RepositoryRegistry) {
         while let Some(()) = self.receiver.recv().await {
             info!("Received sync message");
-
-            info!("Syncing central records...");
-            self.pull_central_records(&repositories).await;
-            info!("Successfully synced central records");
-
-            info!("Integrating central records");
-            self.integrate_central_records(&repositories)
-                .await
-                .expect("Failed to integrate central records");
-            info!("Successfully integrated central records");
-
-            info!("Syncing remote records...");
-            let remote_records = self.pull_remote_records().await;
-            info!("Successfully pulled remote records");
-            info!("Integrating remote records...");
-            self.integrate_remote_records(remote_records);
-            info!("Successfully integrated remote records");
+            info!("Starting sync...");
+            synchroniser.sync(registry).await;
             info!("Finished sync!");
         }
         unreachable!(
             "Sync receiver has stopped listening as channel has closed. Are the senders dead!?"
         );
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        database::{
-            repository::{get_repositories, CentralSyncBufferRepository},
-            schema::CentralSyncBufferRow,
-        },
-        server::data::RepositoryRegistry,
-        util::{
-            configuration,
-            settings::Settings,
-            sync::{
-                get_sync_actors,
-                translation::test_data::{
-                    item::get_test_item_records,
-                    master_list_name_join::get_test_master_list_name_join_records,
-                    name::get_test_name_records,
-                },
-                SyncConnection, SyncReceiverActor, SyncSenderActor,
-            },
-            test_db,
-        },
-    };
-
-    use super::translation::test_data::{
-        check_records_against_database, extract_sync_buffer_rows,
-        master_list::get_test_master_list_records,
-        master_list_line::get_test_master_list_line_records, store::get_test_store_records,
-    };
-
-    #[actix_rt::test]
-    async fn test_integrate_central_records() {
-        let settings: Settings =
-            configuration::get_configuration().expect("Failed to parse configuration settings");
-        let sync_connection = SyncConnection::new(&settings.sync);
-        let (_, sync_receiver): (SyncSenderActor, SyncReceiverActor) =
-            get_sync_actors(sync_connection);
-
-        let settings = test_db::get_test_settings("omsupply-database-integrate_central_records");
-
-        test_db::setup(&settings.database).await;
-        let registry = RepositoryRegistry {
-            repositories: get_repositories(&settings).await,
-        };
-
-        // use test records with cursors that are out of order
-        let mut test_records = Vec::new();
-        test_records.append(&mut get_test_name_records());
-        test_records.append(&mut get_test_item_records());
-        test_records.append(&mut get_test_store_records());
-        test_records.append(&mut get_test_master_list_records());
-        test_records.append(&mut get_test_master_list_name_join_records());
-        test_records.append(&mut get_test_master_list_line_records());
-
-        let central_records: Vec<CentralSyncBufferRow> = extract_sync_buffer_rows(&test_records);
-        let central_sync_buffer_repository: &CentralSyncBufferRepository =
-            registry.get::<CentralSyncBufferRepository>();
-
-        central_sync_buffer_repository
-            .insert_many(&central_records)
-            .await
-            .expect("Failed to insert central sync records into sync buffer");
-
-        sync_receiver
-            .integrate_central_records(&registry)
-            .await
-            .expect("Failed to integrate central records");
-
-        check_records_against_database(&registry, test_records).await;
     }
 }
